+++ fitting.h
@@ -10,10 +10,10 @@
 #include "commath.h"
 
 typedef struct {
-	float *poss;
+	double *poss;
 	int xlen;
 	int *yb;
-	float *fpars; // a1 and a2
+	double *fpars; // a1 and a2
 } Errparams;
 
 double sphesqerr(const gsl_vector *vars, void *params);
@@ -24,8 +24,8 @@ void minerror(double (*f)(const gsl_vector *va, void *params),
 double paraboloid(const double x, const double y, const double a,
 		const double b);
 double sphere(const double x, const double y, const double rad);
-void sphereslope(const float x, const float y, const float rad, float *nrm);
-void parabslope(const float x, const float y, const float f1, const float f2,
-		float *nrm);
+void sphereslope(const double x, const double y, const double rad, double *nrm);
+void parabslope(const double x, const double y, const double f1,
+		const double f2, double *nrm);
 
 #endif /* INC_FITTING_H */
